Add initial Fedora RPM packaging.
This is an initial commit of the spec file to be used for a Fedora RPM
packaging of the controller. The spec relies on Maven to pull in the
JAVA artifacts. Anotehr dependencies RPm will be created to grab those
dependencies.

There is a README in the fedora folder that describes the process to
build the RPM, install and uninstall. The README also details how to
start the controller.

This is a work in progress. I wanted to commit the current spec to let
others take alook. The spec file has comments indicated by
"todo" to show the current thinking of what needs to be added.

Change-Id: Ie93cb57a8012033d0578fa87a4f2cf536d1b0d75
Signed-off-by: Sam Hague <shague@redhat.com>