Add jar dependencies package.
The current opendaylight-controller rpm uses the JAVA artifacts that
are already included in the distribution. This is not allowed in a
proper Fedora distribution. This commit creates a "dependencies"
package to remove the use of the pre-built artifacts from the
controller rpm. In the future this dependencies package will go
away as the artifacts are added in their own packages.

Change-Id: I7cc3891e6f18428b13133d8a0ff999ef1d79158c
Signed-off-by: Sam Hague <shague@redhat.com>