Minor fixups for mock and cloning clean-up
The mock build location repository needs initialized for each run. Each
of the subsequent calls to mock can continue to use the --no-clean
feature of mock but not doing an initial --init causes it to bomb on
Jenkins managed systems.

Clone at --depth 0. The builds are only going to be pulling against the
current master HEAD or a particular version. We may want to change the
clones to be direct fetches to the snapshot tarball interface.

Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>