Add in verbose flag to mock
Jenkins is a bit unhappy about the _very_ large pauses in output that
it's getting when we enter the mock build phase (multiple times). Add
the verbose flag to mock to keep Jenkins happy about this and also make
it easier to track where it is in the build process.

Change-Id: I2c15fc8161511b079eb96178980e2c909a710b91
Signed-off-by: Andrew Grimberg <agrimberg@linuxfoundation.org>