Support multi-RPM builds via common var configs
RPM builds are now defined by build variable config files.

build.sh now accepts a path to the build config you'd like to run, or
defaults to running all build configs defined in the build_vars dir if
no path is given.

The goal here is to support building multiple versions of ODL with the
same version of the codebase. For example, we currently need to be able
to support Helium and Lithium SRs.

Change-Id: Ibe802484e506c21651f61b7210af59469be1231c
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>