Add ODL systemd unitfile to ODL upstream repo
A historical remnant of the RPM's development is that the ODL systemd
unitfile packaged into the RPM is hosted in a stand-alone, external
repo.

https://github.com/dfarrell07/opendaylight-systemd

Moving the unitfile to the proper upstream repo is a obvious win.

Once this is committed, the ODL spec file can be updated with a pointer
to the new source artifact and new RPMs can be cut.

Change-Id: I91122050ea076ac3d395bd4c1a92a15d53e7ffe9
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>