Refactor RPM build to use YAML+Py+Jinja2 templates
While this is a natural iteration on the RPM building logic, it's nearly
a total re-write of the artifacts that embody that logic, which is why
it's so large. Painful, but necessary for it to actually work at all.

The previous approach used RPM macros, read in from .sh vars files and
passed to the .spec file via rpmbuild, to customize the build per-RPM.

This worked well, but turned out be drastically more complex on a remote
build system like the CentOS CBS. Koji requires the macros to be defined
in a painful way (via a second RPM to install the macro default file).

This method reads build variables from a YAML config file, then uses
Jinja2 templates to create static ODL .spec files for each RPM build.
This moves the variable handling logic to Python, before the RPM's
building logic is invoked, avoiding the need to define macros in Koji.

Change-Id: Id12b32c14cab1df3d1580be9c77e1eca25113094
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>