Add tox config for running flake8 tests
This tox configuration file describes how to run flake8, a Python style
checker, against the repository. This will be used for automated verify
jobs.

The config for flake8 includes a 100 char line limit, which is less
strict than PEP8's original 80 but within the 80-100 updated limit.

https://www.python.org/dev/peps/pep-0008/#maximum-line-length

All other PEP8 checks are kept as-is. Additional flake8 config is
focused on providing helpful log output.

Change-Id: I1bd75203966f3713021e00bc3f5d8e17952e123a
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>