Use strings for all RPM version variables
The new Be ERP rpm_version values of the format

<rpm_release>.<ERP date version>

Broke the `-v` param's int expectations, so that argparse option was
removed. Params to build.py then become strings by default. The
build_vars configuration used unquoted numbers for version values, which
were being read in as ints, causing the comparison between version
values passed to the script and those read in from build_vars to fail
based on type differences. Quotes were added to the data to specify
explicitly that version vars should be treated as strings.

Change-Id: I8b20edc7a26db793d084b9b0899c5e5138af4ab9
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>