Add repo config for (future) ODL Be release RPM
This repo config will be used for installing ODL Beryllium, once it's
released. OPNFV's installer and testing projects, as well as ODL
projects like the Ansible role and vagrant-opendaylight, curl down these
example files for use. Adding this before the release allows integration
work to happen in advance. Once the Be RPM is built and tagged to the
nfv7-opendaylight-40-release destination tag, the repo pointed to by
this config will be built and installs will start working.

Change-Id: I13cf676141c1c3e2caf6d605814f4f7aa9db4cb1
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>