Add 4.2.0-1.20160407.144137180 RPM
Builds ODL from a distribution build, instead of the normal autorelease
build. OPNFV needs a fix that's in this build but not the latest
autorelease.

Note that the distribution-karaf-0.4.2-20160407.144137-180.tar.gz
tarball extracts to a distribution-karaf-0.4.2-SNAPSHOT, which breaks
rpmbuild's assumptions about tarballs. It expects the codename to
follow 0.4.2, and the two have different codenames. I'm manually
untaring, renaming and re-taring the cached version to give it
the same dir name as the tarball. This is a hack but this is just going
to the 4-testing CBS repo so it's not a big deal.

Change-Id: I5fb8475f7bec9a5ef8d0a0899932f2a451bf7852
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>