Update patch set 2
Patch Set 2:

> Daniel,
 > 
 > I look at these every time, but it's all just feeling like what I
 > looked
 > at before and I'm slowly becoming numb to the reviews.
 > 
 > Is this just something I need to get used to, or do we find some
 > better way to not have you have to wait on a committer to look at
 > these (just incrementing versions and changing URL) kinds of
 > patches.
 > 
 > I don't mind, but I end up being a slow human in the way of
 > something
 > that should probably be automatic and slowing your work.

The reviews don't really slow me down. For better or worse, I build the SRPM and push to the CBS before this is merged. Tracking the builds here is about documenting how they were created, making them reproducible for me/others later. I could start +2ing and merging myself, to avoid the load of nearly-rote review load on you and Thanh, but I have the impression that we'd prefer to avoid self-merges.

Patch-set: 2