Update patch set 3
Patch Set 3: Code-Review-1

(2 comments)

Congratulations, this produces a working package!

If Daniel agrees, all that remains before merging this particular version is to get the commit message right, and perhaps add a README file (actually, deb/debian/README.Debian so it gets installed in the package). We can go on to packaging 102 in a later patch set...

Patch-set: 3
Label: Code-Review=-1