Update patch set 5
Patch Set 5:

> In the next patch, do we want to change the path to just
 > `integration/packaging/deb/<files here>` or maybe `integration/packaging/deb/opendaylight/<files
 > here>`? This extra `debian` dir seems unnecessary to me.

All the Debian packaging tools expect the Debian-specific files to be in a "debian" sub-directory of the directory containing the files to be packaged, so we need to keep the "debian" directory. Given the packaging structure used in the "rpm" directory, it might well make sense to move the "debian" directory to an "opendaylight" directory (to differentiate the support infrastructure, Vagrantfiles etc., which don't belong in the package).

Thus we'd end up with

deb/build.sh
deb/README.md
deb/Vagrantfile
deb/opendaylight/debian/changelog
deb/opendaylight/debian/...

(assuming we'll end up with a Vagrantfile and build.sh script of course).

Patch-set: 5