Add Vagrantfile for L2Switch tutorial
Defines a Vagrant env with OpenDaylight, OVS and Mininet, with ODL's
L2Switch and DLUX features installed and DLUX's port forwarded to the
host.

A user should be able to `vagrant up` this box, wait for provisioning,
connect to `localhost:8181/index.html`, login with `admin`/`admin` and
see ODL DLUX using L2Switch to show a Mininet topology.

Change-Id: Ifb717d64aedd62be066934aeb22339ce9fb5791e
Signed-off-by: Girish Sukhatankar <girish.sukhatankar@colorado.edu>
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>