Update patch set 3
Patch Set 3:

It'd be super helpful to define a Vagrantfile that creates an env for building the debs, and add docs about exactly how to build them, so I can more easily test these. We'll want that Vagrantfile eventually anyway, as a part of making the build reproducible by all.

Patch-set: 3