Update patch set 7
Patch Set 7:

It's really strange. I prevented packer from removing the linuxaddition iso that was provided in /hom/vagrant. I logged in to the VM that packer uses to create the image, and I tried to install manually but impossible. I can't figure out why, but looks like it cannot extract properly, or something's not compatible.

Patch-set: 7