Set expected ODL exit code to 143 in RPM unitfile
ODL's normal exit value seems to be 143. Tell systemd that this isn't an
error, but the code it should expect ODL to return. Will cause systemd
to start reporting ODL exits as successful.

Once this is merged, new releases for all current RPMs should be cut
with sysd_commit set to find this unitfile version.

Based on 40584 by Akshita Jha.

Change-Id: I95a18a58f6201d2e76402770557a7b90245830cd
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>