Update patch set 9
Patch Set 9:

> > > Also, the autorelease build (SR3) package uses 0 as
 > major_version
 > > > (vs shifting 4 to major, like the 4.2 build and RPMs do).
 > >
 > > Actually, the 4.2 build does this too. imho, we should shift the
 > > versions to the left of what they are in the tarballs.
 > 
 > I am sorry but I don't clearly understand what this means? Does it
 > mean that we should change the .deb pkg name to something like
 > opendaylight_<version_major>.<version_minor>.<version_patch>-<pkg_version>_amd64.deb
 > instead of the current opendaylight_0.<version_major>.<version_minor>-<codename>_amd64.deb
 > ?

Yes, exactly. The discussions I've heard around the future of ODL versioning seem to agree on using the ODL major release for the major version field (Li=3.0.0, Be=4.0.0, Be SR3=5.3.0, B=5.0.0 and similar). I need to write some docs about versioning for another TODO, but haven't had the cycles. Will bump+ prio and link you to it when done.

Patch-set: 9