Update patch set 1
Patch Set 1: Code-Review-1

(3 comments)

Hello Patricio,

Thanks for the patch.
 
 > Daniel, you are right the main two objectives are
 > a) Generate from a local tar.gz (usually coming from a just
 > generated compillation of distribution)

This is a good objective. Sounds reasonable.

 > b) Keep it simple in order to not require other software packages.

This can prove beneficial, but as Daniel already mentioned since you are not using templates and YAML file, how do you plan to account for the changelog time/date, Java version (Java 7 for Beryllium and Java 8 for Boron) etc. ?

 > Since I started this a couple of days before Akshita python build
 > script, in fact last night i had to rebase when i pulled the lates
 > code,  i based everything on bash, however since now there is a
 > build foundation, i can work on understanding the actual python
 > build process and add the required elements to perform the local
 > build operation based on the python existing scripts. 

So are you planning to add this local build logic to the existing python script (build.py)? If not, I can go ahead with that. Do let me know.

 > Anyways i
 > will perform a quick fix of the elements mentioned by Stephen and
 > rebasing it to packaging/scripts so people can take advantage of
 > this and in the mean time work on using the new python process.
 > Does this make sense?

Sure. I don't think there is any harm doing this.

Patch-set: 1