Update patch set 1
Patch Set 1:

The other thing that needs to be done is to build the package on the CentOS CBS and tag it to these repos. If you want to do that, we can get you setup with a CBS account and committer rights. These are the (lacking) docs: https://wiki.centos.org/HowTos/CentosPackager https://wiki.centos.org/HowTos/CommunityBuildSystem

You used to have to file a bug like this to get access, but I guess they have changed that because it's no longer in the docs above.

https://bugs.centos.org/view.php?id=8879

Patch-set: 1