Update patch set 1
Patch Set 1:

Also unrelated to this change, I noticed that the Vagrant env is failing to provision with 404s about broken URLs during `apt-get install`. We need to `apt-get update` to get new package source info. I have the change locally already, I'll send a patch and add you as a reviewer in a moment.

Patch-set: 1