Add Packer config make ssh auth work by default
Recent versions of Fedora don't use ssh-dss, which old versions of ODL
seem to offer by default. Can fix with a system config change like this
or by manually passing as option to ssh command. The csit and l2switch
tutorials do it by changing this configuration. vagrant-opendaylight's
connect.sh helper script is an example of passing the param to ssh CLI.

There might be a better way/place to fix this. Ideas encouraged.

Change-Id: Id143fa6bc864469926596dfaddb77902f6d8c68f
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>