Update patch set 14
Patch Set 14:

Said on an email, but adding here for clarity:

We shouldn't remove the CLI flags for passing version info. They can be optional, overriding the version info automatically parsed from the URL if they are passed. This is important for use cases like building the latest snapshot from a given branch, when we just want to pass major version and not a URL.

https://trello.com/c/5Uvaf3R4/397-add-rpm-build-py-logic-to-build-latest-snapshots

Patch-set: 14