Update patch set 28
Patch Set 28:

(1 comment)

To answer Daniel's question it should be the individual eg. Girish Sukhatankar, or the Organization they are representing who contributed the majority of the existing lines of code in the file. This should get updated when another individual/org becomes the entity who contributed the most significant amount of current code to the file.

If you're interning for OpenDaylight setting this to "OpenDaylight" is fine too.

Patch-set: 28
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0