Update patch set 2
Patch Set 2:

I guess there are three use-cases:

1) Build package defined in build_vars.yaml
2) Build a new package, passing required params.
3) Build the latest snapshot from the given branch (major/minor tuple)

For 1, I like -v because it's easy for humans.

For 2, we mostly plan on it being used by automated systems with easy access to URL, so wanted to centralize extracting info from URL in build.py vs having to add version-extracting logic to places that call this.

Three is meant for both humans and automated systems, can find the URL on its own.

Patch-set: 2