Update patch set 1
Patch Set 1: Code-Review-1

(8 comments)

Thanks for pushing what you have so far Alok. :)

In addition to the comments inline, Jenkins' verify job has some feedback.

Now that you have this in a public Gerrit, I suggest working on updating the RelEng/Builder jobs to consume this new versions. We'll need to have them working in the sandbox with the refactored logic before merging this of course.

Patch-set: 1
Label: Code-Review=-1