Update patch set 3
Patch Set 3:

> (1 comment)

I did it, but I kinda don't like the result. I don't see a way to use custom link text with this method. Seems I either get an ugly link with the same name as the ref or the next section title, which doesn't fit cleanly in the places I'm trying to us it.

https://thomas-cokelaer.info/tutorials/sphinx/rest_syntax.html#explicit-links

Patch-set: 3
Label: Verified=0