Update patch set 6
Patch Set 6:

Ah yes. The errors are because rstlintbear does not support Sphinx and is incompatible with it. There are supposibly ways to add ignore rules but there's so many sphinx directives that it would be a large task to catch them all I think...

In the docs project we used the RSTCheckBear instead which is more generic (doesn't check directives) but at least works with Sphinx.

Patch-set: 6
Reviewer: Gerrit User 2759 <2759@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0