Update patch set 4
Patch Set 4:

> -1 until I see a working sandbox run

I am unable to get a working sandbox run of build-rpm and build-rpm-snap jobs. The script gets executed fine in both cases but the part which deals with uploading the artifact is failing I guess.
 
Here I ran the job with gerrit-refspec refs/changes/08/60308/4
https://logs.opendaylight.org/sandbox/jenkins091/packaging-build-rpm-carbon/4/console.log.gz
and this is a normal run.
https://logs.opendaylight.org/sandbox/jenkins091/packaging-build-rpm-carbon/3/console.log.gz

and the same is the case with build-rpm-snap
Link for job run with gerrit refspec refs/changes/08/60308/4
https://logs.opendaylight.org/sandbox/jenkins091/packaging-build-rpm-snap-carbon/4/console.log.gz
and this is a normal run
https://logs.opendaylight.org/sandbox/jenkins091/packaging-build-rpm-snap-carbon/5/console.log.gz

Can you please consider this?

Patch-set: 4