Add example RPM repo configs for ODL Nexus repos
Add .repo files that will install RPMs from each of the new RPM repos
hosted on ODL's Nexus.

Not adding .repo files for release repos yet as there are no RPMs in
them and trying to install from them errors out.

Not adding 7-devel repo because there are no Nitrogen builds in it yet
and trying to install from it errors out.

Change-Id: Id09048c837d80bf197dde8f2b1fa7ba25c8e93cb
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>