Update patch set 3
Patch Set 3:

> Daniel, can you remove your -2 please and give +2 :P

No, not yet, I need to see it work with the upgrade test job. I'll work on this.

We also have a huge set of changes queued up to refactor the packaging logic that I'd really like to land, then rebased this on top of (vs trying to rebase all of those on top of this).

https://git.opendaylight.org/gerrit/#/q/status:open+project:integration/packaging+branch:master+topic:refactoring

Patch-set: 3