Update patch set 6
Patch Set 6:

> > Would this still be helpful for deployments without Director?
 > 
 > Have explained why I think it would not really in private email
 > reply.

And I agree. Here's the explanation for docs:

there are at least 2 problems:

a) this won't be enough for deployments that don't use Director.. there are other steps we need. The intention is that, once we have it all work, one fine day, we would write up everything we automated in director so that, in theory (!), someone could manually do the same.

b) even if you wanted to this for "at least one part of the story" line of thought, then IMHO it creates the "impression" that the journal/ snapshot/ etc/ directories are guaranteed to be compatible accross upgrades. However that is not so! That's what the whole update VS upgrade and level 1 VS 2/3 thing we're doing is all about. So while this MAY work for some updates (minor, level 1) it won't for at least some major upgrades (which is why Josh is doing all this work for replay and we'll orchestrate from Director; this is documented in more details in our Upgrade-abilty Google Doc).

Long story short, I have a feeling that by putting this into the RPM Spec it may create more confusion that it helps anyone. We don't need it downstream anymore, so my suggestion would be to forget about it.

Patch-set: 6