Update patch set 2
Patch Set 2: Code-Review+2

Yeah, I think this will work when actually merged and it's just a limitation of how the verify job works. I don't have a good idea for how to fix the verify job as it really does need a merged hash for the unitfile version and it of course should use the under-dev code. When changing the unitfile, it can't verify that part of the change. I guess we could give up on using a known version of the unitfile and just use whatever's in the repo.

I'll test this after merge and revert if it's failing.

Patch-set: 2
Label: Code-Review=+2