Update patch set 1
Patch Set 1:

> Thanks for being such a great "Do-er" Thanh;)

NP. Integration/Packaging is now available at this URL:

http://docs.opendaylight.org/projects/integration-packaging/

I'll look into linking it from the docs project so that it appears on the landing page too. If you have a RTD account ping me with your username privately on IRC and I can add you to the management for the RTD project.

Patch-set: 1