Update patch set 2
Patch Set 2:

> We should probably look into adding mocking tests so that the URLs
 > do not need to actually exist. I think the important thing to test
 > here is simply that the URL patterns we are expecting is parseable
 > by the code.

Yeah, I've thought about this but I'm not sure how to implement it, or if it's possible. The issue is that the logic under test curls down HTML it finds from the URL to extract data like the date of the build, which it uses in pre-release version numbers (it extracts other things too I think but I'd need to dig into the code to refresh myself). So it seems like we really do need real, valid URLs to test that logic.

Patch-set: 2