Add override for pkg version in pkg build logic
The package version is the bit after the minus, meant for use by the
package itself, not the thing being packaged (ODL). It's normally
auto-detected for pre-release builds, or manually fixed at 1 for
releases. Add param to optionally manually override.

This is needed to fix an issue with building the Fluorine Common Distro
on the CentOS CBS.

Jira: INTPAK-210
Change-Id: Ifc97d92d4ec3503b31ea57eb535cdf79505afcf0
Signed-off-by: Daniel Farrell <dfarrell@redhat.com>